Support `inherit_gem` directive
This commit makes rubocop.yml which contains `inherit_gem` directive
compatible with this gem and the custom action on top of it.

Currently the config which contains these lines does not work:

```
inherit_from: .rubocop_todo.yml

inherit_gem:
  rubocop-rails-omakase: rubocop.yml
```

This is the workflow definition which we have attempted;

```
name: rubocop-todo-corrector

on:
  pull_request:
    types:
      - closed
  workflow_dispatch:
    inputs:
      cop_name:
        description: Pass cop name if you want to pick a specific cop.
        required: false
        type: string

jobs:
  run:
    runs-on: ubuntu-latest
    permissions:
      contents: write
      pull-requests: write
    steps:
      - uses: r7kamura/rubocop-todo-corrector@v0
        with:
          cop_name: ${{ inputs.cop_name }}
          github_token: ${{ secrets.GITHUB_TOKEN }}
          label: rubocop-todo-corrector
```

and this is the error it produced:

```
Unable to find gem rubocop-rails-omakase; is the gem installed? Gem::MissingSpecError
```