Employ non-transient Gemfile when testing describe command
The previous commit has had to change spec/fixtures/dummy_rubocop.yml,
which turns out to be depended on by this spec that needs to be fixed.

Before this commit, this test subject depended on tmp/Gemfile_rubocop_todo_corrector.rb
which was ephemeral and test-friendly in appearance, but it was a byproduct of
another independent test of Commands::Bundle. Therefore attempting to run the unit test
without the presence of the tmp file was an inevitable failure:

```
rm -rf tmp && bundle exec rspec spec/rubocop_todo_corrector/commands/describe_spec.rb
=> /tmp/Gemfile_rubocop_todo_corrector.rb not found (Bundler::GemfileNotFound)
```

In fact, making a compromise and depending on the constant ordering of test cases does not help:

```
rm -rf tmp && bundle exec rspec
=> Could not find gem 'rubocop-rails-omakase' in locally installed gems. (Bundler::GemNotFound)
```

It is because the change to spec/fixtures/dummy_rubocop.yml to include
inherit_gem in the previous commit has changed the nature of the temporal Gemfile
since `rubocop-rails-omakase` (or any other gem which can be depended on through
inherit_gem) is none of the dependencies of this library. It is expected to be absent,
therefore Bundler is right that it's unable to find the gem in the local.

To better illustrate the situation described by this point, this is the expected content
of tmp/Gemfile_rubocop_todo_corrector.rb after finishing commands/bundle_spec.rb;

```
source 'https://rubygems.org'
gem 'rubocop', '1.29.1'
gem 'rubocop-rake', '0.6.0'
gem 'rubocop-rspec', '2.10.0'
gem 'rubocop-rails-omakase'
```

and the source of the trouble is at CopSourceDetector where it tries to `#capture`
in which a command like this is expected to run under the hood, and a message like this is to pop up;

```
BUNDLE_GEMFILE=tmp/Gemfile_rubocop_todo_corrector.rb bundle exec ruby -e 'Bundler.require'
=> Could not find gem 'rubocop-rails-omakase' in locally installed gems. (Bundler::GemNotFound)
```

This is when I believe replacing tmp/Gemfile_rubocop_todo_corrector.rb with
the regular Gemfile could be better for this test subject; the reasons being;

1. Gemfile, unlike tmp/Gemfile_rubocop_todo_corrector.rb, is safer to expect the presence
    so it lays a better protection against Bundler::GemfileNotFound, and
1. tmp/Gemfile_rubocop_todo_corrector.rb since the previous commit is no longer a representation of
    locally installed gems which are safely require-able, while Gemfile remains to guarantee
    that its contents are in the local which are safe to require. In short, using Gemfile
    is less prone to Bundler::GemNotFound.